fix T5Tokenizer

+++ app/core/config.py
@@ -23,7 +23,7 @@ class Settings(BaseSettings):
     mt_mock: bool = False
     mt_model_id: str = "google/madlad400-3b-mt"
     mt_device_map: str = "auto"
-    mt_dtype: str = "bfloat16"
+    mt_dtype: str = "float16"
     # After each job, park weights in RAM so the 12GB card is free for LTX/TTS/Comfy.
     # Translate still runs on GPU. Set 0 only if MT owns the GPU.
     mt_free_vram: bool = True
